+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added cockroach_backups data source for listing backups on a specific cluster.
+- Added cockroach_restores data source for listing restores on a specific cluster.
 
 ### Fixed
 - Ensure all cluster regions are ready when waiting or private endpoint